cody-gateway: prepare client to support both `int` and `BigInt` for rate limit (#53289)
This PR makes the client (Cody Gateway) compatible with the current
`int` and (future) `BigInt/string` types for rate limit quota by
temporarily unmarshaling it to `any` and then performing the type
assertion.

The compatibility is mainly for a smooth initial rollout and will be
changed to deal with `BigInt` after the initial rollout specifically.

## Test plan

Manually tested with the old and new dotcom backend (in
https://github.com/sourcegraph/sourcegraph/pull/53290), that both
scenarios correctly updated the rate limit cache in the Redis.

---

Part of https://github.com/sourcegraph/sourcegraph/issues/53007

+++ enterprise/cmd/cody-gateway/internal/actor/dotcomuser/dotcomuser.go
@@ -142,7 +142,7 @@ func newActor(source *Source, cacheKey string, user dotcom.DotcomUserState, conc
 
 	if rl := user.CodyGatewayAccess.ChatCompletionsRateLimit; rl != nil {
 		a.RateLimits[codygateway.FeatureChatCompletions] = actor.NewRateLimitWithPercentageConcurrency(
-			rl.Limit,
+			int64(rl.Limit),
 			time.Duration(rl.IntervalSeconds)*time.Second,
 			rl.AllowedModels,
 			concurrencyConfig,
@@ -151,7 +151,7 @@ func newActor(source *Source, cacheKey string, user dotcom.DotcomUserState, conc
 
 	if rl := user.CodyGatewayAccess.CodeCompletionsRateLimit; rl != nil {
 		a.RateLimits[codygateway.FeatureCodeCompletions] = actor.NewRateLimitWithPercentageConcurrency(
-			rl.Limit,
+			int64(rl.Limit),
 			time.Duration(rl.IntervalSeconds)*time.Second,
 			rl.AllowedModels,
 			concurrencyConfig,
@@ -160,7 +160,7 @@ func newActor(source *Source, cacheKey string, user dotcom.DotcomUserState, conc
 
 	if rl := user.CodyGatewayAccess.EmbeddingsRateLimit; rl != nil {
 		a.RateLimits[codygateway.FeatureEmbeddings] = actor.NewRateLimitWithPercentageConcurrency(
-			rl.Limit,
+			int64(rl.Limit),
 			time.Duration(rl.IntervalSeconds)*time.Second,
 			rl.AllowedModels,
 			concurrencyConfig,
